+++ packages/calendar-range/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @alfalab/core-components-calendar-range
 
+## 8.0.9
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-calendar@8.0.8
+    - @alfalab/core-components-calendar-input@11.0.9
+
+## 8.0.8
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-calendar@8.0.7
+    - @alfalab/core-components-calendar-input@11.0.8
+
 ## 8.0.7
 
 ### Patch Changes