+++ packages/date-range-input/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @alfalab/core-components-date-range-input
 
+## 4.0.9
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-calendar@8.0.8
+
+## 4.0.8
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-calendar@8.0.7
+
 ## 4.0.7
 
 ### Patch Changes