+++ packages/date-time-input/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @alfalab/core-components-date-time-input
 
+## 5.0.9
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-calendar@8.0.8
+
+## 5.0.8
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-calendar@8.0.7
+
 ## 5.0.7
 
 ### Patch Changes