+++ packages/drawer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @alfalab/core-components-drawer
 
+## 6.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-base-modal@6.1.0
+
 ## 6.0.2
 
 ### Patch Changes