+++ packages/gallery/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @alfalab/core-components-gallery
 
+## 6.0.11
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-tooltip@8.0.10
+
+## 6.0.10
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-base-modal@6.1.0
+    - @alfalab/core-components-tooltip@8.0.9
+
 ## 6.0.9
 
 ### Patch Changes