+++ packages/international-phone-input/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @alfalab/core-components-international-phone-input
 
+## 3.2.7
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-select@18.2.7
+    - @alfalab/core-components-input-autocomplete@13.0.10
+
+## 3.2.6
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-select@18.2.6
+    - @alfalab/core-components-input-autocomplete@13.0.9
+
 ## 3.2.5
 
 ### Patch Changes