+++ packages/picker-button/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @alfalab/core-components-picker-button
 
+## 12.0.10
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-select@18.2.7
+
+## 12.0.9
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-select@18.2.6
+
 ## 12.0.8
 
 ### Patch Changes